Separate herd form into own file
